fixed logic error in if else

+++ src/filesys.c
@@ -295,7 +295,7 @@ void remove_directories(int img_fd, bpb_t bpb, const char* dir_name) {
             // Remove file or recursively remove directory
             if (dirEntry->DIR_Attr & 0x10 && strcmp(dirEntry->DIR_Name, ".") != 0 && strcmp(dirEntry->DIR_Name, "..") != 0) { // Directory
                 remove_directories(img_fd, bpb, entryName);
-            } else { // File
+            } else if(strcmp(dirEntry->DIR_Name, ".") != 0 && strcmp(dirEntry->DIR_Name, "..") != 0){ // File
                 remove_file(img_fd, bpb, entryName);
             }
         }
